Fixing hdf5 compilation on osx
Former-commit-id: 011f69f51b1e489e342a26846878867a48ff11a8

+++ source/algorithms/class_base_algorithm.cpp
@@ -441,7 +441,7 @@ int class_base_algorithm::env_storage_move(){
 
 void class_base_algorithm::print_status_update() {
     if (Math::mod(iteration, print_freq) != 0) {return;}
-    if ((position) != superblock->environment_size/2){return;}
+    if ((unsigned long)position != superblock->environment_size/2ul){return;}
     if (print_freq == 0) {return;}
 
     compute_observables();
